+++ services/avery/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Stop windows service on runtime errors correctly.
+
 ## [1.1.0] - 2021-09-09
 
 ### Added